backend/app/utils/clerk.py
import jwt
from typing import Optional
from fastapi import HTTPException, status
from app.config import settings


def verify_clerk_token(token: str) -> dict:
    """
    Verify a Clerk JWT token and return the payload.
    
    Args:
        token: The JWT token from the Authorization header
    
    Returns:
        Decoded token payload with user information
    
    Raises:
        HTTPException if token is invalid
    """
    try:
        # Remove "Bearer " prefix if present
        if token.startswith("Bearer "):
            token = token[7:]
        
        # Get Clerk's public key (in production, fetch from Clerk's JWKS endpoint)
        # For now, we'll verify using the secret key
        # Note: Clerk uses RS256, so we need the public key
        # This is a simplified version - in production, fetch from Clerk's JWKS
        
        # Decode the token
        # Clerk tokens are signed with RS256, so we need the public key
        # For MVP, we can use Clerk's backend API to verify
        # Or use the secret key if available
        
        # Simplified verification - in production, use Clerk's verification endpoint
        # or fetch the public key from Clerk's JWKS endpoint
        decoded = jwt.decode(
            token,
            settings.CLERK_SECRET_KEY,
            algorithms=["HS256"],  # Clerk uses RS256, but for MVP we'll handle this differently
            options={"verify_signature": False}  # For MVP, we'll verify via Clerk API
        )
        
        # Verify the token with Clerk's backend API
        # For now, we'll do basic validation
        # In production, make a request to Clerk's verification endpoint
        
        return decoded
        
    except jwt.ExpiredSignatureError:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Token has expired"
        )
    except jwt.InvalidTokenError:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid token"
        )
    except Exception as e:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail=f"Token verification failed: {str(e)}"
        )


def get_clerk_user_id(token: str) -> str:
    """
    Extract the Clerk user ID from a verified token.
    
    Args:
        token: The JWT token
    
    Returns:
        Clerk user ID
    """
    payload = verify_clerk_token(token)
    # Clerk user ID is typically in the 'sub' claim
    user_id = payload.get("sub") or payload.get("user_id")
    if not user_id:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Token does not contain user ID"
        )
    return user_id
